+++ betalm/src/main/java/eu/monnetproject/translation/langmodels/smoothing/NGramHistories.java
@@ -41,7 +41,7 @@ public interface NGramHistories {
      */
     int N();
     
-    Object2ObjectMap<NGram,double[]> histories(int n);
+    Object2ObjectMap<NGram,float[]> histories(int n);
     
     int[][] countOfCounts();
 }
